Fix undefined this by using arrow function
I tell you why. An arrow function does not newly define its own this when it's being executed in the global context; instead, the this value of the enclosing lexical context is used, equivalent to treating this as closure value. The arrow function this is the same as the linked context.